nowcasting_inla

function to estimate amount of events already started by not yet notified. The main use is to estimate how many cases in a outbreak has already started their onset date of symptons but has not yet notified. nowcasting_inla, fits a statistical distribution to the empirical distribution of time of delay between onset date and report date.

Usage

nowcasting_inla(
  dataset,
  bins_age = "SI-PNI",
  trim.data = 0,
  Dmax = 15,
  wdw = 30,
  use.epiweek = FALSE,
  age_col,
  date_onset,
  date_report,
  data.by.week = FALSE,
  silent = F,
  K = 0,
  trajectories = F,
  zero_inflated = F,
  ...
)

Arguments

dataset

Dataset with at least 2 columns, date of onset, date of report. It can be a dataset with 3 columns, two dates columns as before said and a another one being an stratum column, in which data will stratified, usually being age.

bins_age

Age bins to do the nowcasting, it receive a vector of age bins, or options between, "SI-PNI", "10 years", "5 years". Default "SI-PNI".

trim.data

(in weeks) Date to be trimmed out from the data base, in days. Default 0 days.

Dmax

(in weeks) Window of dates the estimation will act, i.e., till how many past weeks the nowcasting will estimate. Default 15 weeks.

wdw

(in weeks) Until which maximum amount of weeks the Nowcasting will use to the estimation. Default 30 weeks.

use.epiweek

If TRUE, it uses the CDC epiweek definition where the week starts on Sunday, if FALSE it the week ends at the weekday of the last record date. Default FALSE

age_col

Column for ages

date_onset

Column of dates of onset of the events, normally date of onset of first symptoms of cases

date_report

Column of dates of report of the event, normally date of digitation of the notification of cases

data.by.week

If it has to be returned the whole time-series data. Default FALSE.

silent

Deprecated Should be the warnings turned off? Default is TRUE.

K

(in weeks) How much weeks to forecast ahead? Default K is 0, no forecasting ahead

trajectories

Returns the trajectories estimated from the inner 'INLA' model Default FALSE.

zero_inflated

Experimental In non-structured models, fit a model that deals with zero-inflated data. Default FALSE. If the age_col is not missing this flag is ignored.

...

list parameters to other functions

return.age

Deprecated If the estimate by Age should be returned. Default TRUE.

Value

a list of 2 elements, each element with a data.frame with nowcasting estimation, 'Total', 'data' with the time-series out of wdw . If 'age_col' is parsed, add a thrid element with by age estimation 'age' . If 'trajectories' = TRUE, add a forth element with the returned trajectories from 'inla'.

Examples

# Loading Belo Horizonte SARI dataset
data(sragBH)
nowcasting_inla(dataset = sragBH,
                date_onset = DT_SIN_PRI,
                date_report = DT_DIGITA,
                silent = T)
#> Joining with `by = join_by(date_onset)`
#> Joining with `by = join_by(date_onset, delay, Time)`
#> $total
#> # A tibble: 15 × 7
#>     Time dt_event   Median    LI    LS   LIb   LSb
#>    <int> <date>      <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1    17 2021-12-13   443   442   449   443   445 
#>  2    18 2021-12-20   632   626   641   629   635 
#>  3    19 2021-12-27   736   727   751   733   741 
#>  4    20 2022-01-03   760   745   778.  754   765 
#>  5    21 2022-01-10   879   862   904   872   887 
#>  6    22 2022-01-17   786   763   815   777   795 
#>  7    23 2022-01-24   590   564   624   580   600 
#>  8    24 2022-01-31   511   483   550.  500   522 
#>  9    25 2022-02-07   408   377   454.  397   422 
#> 10    26 2022-02-14   404   365   454.  390   420 
#> 11    27 2022-02-21   337   294   399   320   359 
#> 12    28 2022-02-28   300.  246.  379.  279.  323 
#> 13    29 2022-03-07   311   245.  424   289   347 
#> 14    30 2022-03-14   292   197   448.  250   334.
#> 15    31 2022-03-21   248   129.  501.  201   319.
#>