nowcasting_inla

function to estimate amount of events already started by not yet notified. The main use is to estimate how many cases in a outbreak has already started their onset date of symptons but has not yet notified. nowcasting_inla, fits a statistical distribution to the empirical distribution of time of delay between onset date and report date.

Usage

nowcasting_inla(
  dataset,
  bins_age = "SI-PNI",
  trim.data,
  Dmax = 15,
  wdw = 30,
  age_col,
  date_onset,
  date_report,
  data.by.week = FALSE,
  return.age = NULL,
  silent = F,
  K = 0,
  trajectories = F,
  ...
)

Arguments

dataset

Dataset with at least 3 columns, date of onset, date of report and stratum.

bins_age

Age bins to do the nowcasting, it receive a vector of age bins, or options between, "SI-PNI", "10 years", "5 years". Default "SI-PNI".

trim.data

(in days) Date to be trimmed out from the data base, in days.

Default 0 days.

Dmax

(in weeks) Until which maximum amount of weeks the Nowcasting will use for the estimation. Default 15 weeks.

wdw

(in weeks) Window of dates the estimation will act, i.e., how many past weeks the nowcasting will estimates. Default 30 weeks.

age_col

Column for ages

date_onset

Column of dates of onset of the events, normally date of onset of first symptoms of cases

date_report

Column of dates of report of the event, normally date of digitation of the notification of cases

data.by.week

If it has to be returned the whole time-series data. Default FALSE.

return.age

Depracted If the estimate by Age should be returned. Default TRUE.

silent

Depracted Should be the warnings turned off? Default is TRUE.

K

How much weeks to forecast ahead? Default K is 0, no forecasting ahead

trajectories

Value

a list of 2 elements, each element with a data.frame with nowcasting estimation, $1 'Total', $2 by 'Age'. If data.by.week = TRUE, add a $3 'dados' with the time-series out of wdw.